THE ACTUAL FINAL GAME
The journey has officially ended, at least for some time. It's now a full functional game with many, many things which was introduced. Objects like health potion and power attack potion, cannon that shoots when the player is in sight, 3 whole new levels, a death screen, an option menu, and power attackkk!!